Rename file to lower-case for neatness.
