+++ src/pages/admin/index.tsx
@@ -2,7 +2,7 @@ import { type NextPage } from "next";
 import React, { useState, useContext } from "react";
 
 import { BestModsPage } from '../../components/main';
-import HeadInfo from "../../components/Head";
+import HeadInfo from "../../components/head";
 import { SessionCtx } from "../../components/main";
 
 import { AlertForm } from '../../components/alert';
