+++ src/pages/index.tsx
@@ -2,7 +2,7 @@ import { type GetServerSidePropsContext, type NextPage } from "next";
 import React from "react";
 
 import { BestModsPage } from '../components/main';
-import HeadInfo from "../components/Head";
+import HeadInfo from "../components/head";
 
 import ModBrowser from '../components/modbrowser';
